résolution erreur 3

+++ api/controller/appartement.php
@@ -1,6 +1,6 @@
 <?php 
 include_once "./service/appartementService.php";
-include_once "./class/appartementUser.php";
+include_once "./class/appartement.php";
 
 
 class AppartementController {
@@ -44,6 +44,9 @@ class AppartementController {
     }
 
     function create_appartement($req, $res) {
+
+
+
         // Récupérer le contenu JSON de la requête
         $json = file_get_contents('php://input');
 
@@ -57,7 +60,9 @@ class AppartementController {
             $res->content = json_encode(array('error' => 'Invalid JSON data'));
             return;
         }
+            //var_dump($data['token']);
 
+        if($this->get_right($data['token'])==1){
         // Créer un nouvel objet Appartement en utilisant les données JSON
         $appartement_object = new Appartement();
         $appartement_object->superficie = $data['superficie'];
@@ -73,6 +78,11 @@ class AppartementController {
 
         // Envoyer une réponse au client
         $res->content = json_encode($new_appartement);
+        }
+        else{
+            $res->status = 401;
+            $res->content = '{"message":"Not enough permission"}';
+        }
     }
 
 
@@ -92,46 +102,70 @@ class AppartementController {
     }
 
     function delete_appartement($req, $res) {
-        if (!isset($req->uri[3])) {
-            $res->status = 400;
-            $res->content = '{"message":"Cannot delete without ID"}';
-        }
-
-        $this->service->delete_appartement($req->uri[3]);
-    }
-
-    function update_appartement($req, $res) {
-        if (!isset($req->uri[3])) {
-            $res->status = 400;
-            $res->content = '{"message":"Cannot update without ID"}';
-        }
 
         $json = file_get_contents('php://input');
-
-        // Décoder le JSON en un tableau associatif
         $data = json_decode($json, true);
-
-        // Vérifier si les données JSON sont valides
         if (!$data) {
-            // Gérer l'erreur de données JSON malformées
-            // Par exemple, renvoyer une réponse d'erreur JSON au client
             $res->content = json_encode(array('error' => 'Invalid JSON data'));
             return;
         }
 
-        // Créer un nouvel objet Appartement en utilisant les données JSON
-        $appartement_object = new Appartement();
-        $appartement_object->superficie = $data['superficie'];
-        $appartement_object->nb_occupant = $data['nb_occupant'];
-        $appartement_object->rue = $data['rue'];
-        $appartement_object->ville = $data['ville'];
-        $appartement_object->cp = $data['cp'];
-        $appartement_object->prix = $data['prix'];
-        $appartement_object->proprietaire = $data['proprietaire'];
+        if($this->get_right($data['token'])==1) {
+            if (!isset($req->uri[3])) {
+                $res->status = 400;
+                $res->content = '{"message":"Cannot delete without ID"}';
+            }
 
-        $new_appartement = $this->service->update_appartement($req->uri[3], $appartement_object);
+            $this->service->delete_appartement($req->uri[3]);
+        }
     }
 
+    function update_appartement($req, $res) {
+        $json = file_get_contents('php://input');
+        $data = json_decode($json, true);
+        if (!$data) {
+            $res->content = json_encode(array('error' => 'Invalid JSON data'));
+            return;
+        }
+
+            if($this->get_right($data['token'])==1) {
+                if (!isset($req->uri[3])) {
+                    $res->status = 400;
+                    $res->content = '{"message":"Cannot update without ID"}';
+                }
+
+                $json = file_get_contents('php://input');
+
+                // Décoder le JSON en un tableau associatif
+                $data = json_decode($json, true);
+
+                // Vérifier si les données JSON sont valides
+                if (!$data) {
+                    // Gérer l'erreur de données JSON malformées
+                    // Par exemple, renvoyer une réponse d'erreur JSON au client
+                    $res->content = json_encode(array('error' => 'Invalid JSON data'));
+                    return;
+                }
+
+                // Créer un nouvel objet Appartement en utilisant les données JSON
+                $appartement_object = new Appartement();
+                $appartement_object->superficie = $data['superficie'];
+                $appartement_object->nb_occupant = $data['nb_occupant'];
+                $appartement_object->rue = $data['rue'];
+                $appartement_object->ville = $data['ville'];
+                $appartement_object->cp = $data['cp'];
+                $appartement_object->prix = $data['prix'];
+                $appartement_object->proprietaire = $data['proprietaire'];
+
+                $new_appartement = $this->service->update_appartement($req->uri[3], $appartement_object);
+            }
+        }
+
+    function get_right($token)
+    {
+       $droit = $this->service->get_right($token);
+        return $droit;
+    }
 
 }
